requirements.txt
google-generativeai
fastapi
uvicorn
pydantic
sqlalchemy
pillow
python-dotenv
